requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn>=0.29.0,<1.0.0
openai>=1.40.0,<2.0.0
chromadb>=0.5.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.0.0,<3.0.0
pymupdf>=1.24.0,<2.0.0
camelot-py[cv]>=0.11.0,<1.0.0
